+++ test.py
@@ -1,9 +1,11 @@
 from that_depends import BaseContainer, container_context, providers
+from that_depends.meta import BaseContainerMeta
 
 
 class Container(BaseContainer):
     """Container for the application."""
 
+    alias = "TEXT"
     default_scope = None
 
     resource = providers.Factory(lambda: 42)
@@ -12,6 +14,8 @@ class Container(BaseContainer):
 def main() -> None:
     with container_context(Container):
         pass
+
+    print(Container.get_providers())
 
 
 if __name__ == "__main__":
